[Refactor] SessionWeb navigation 경계 분리
