+++ Modules/FeatureSession/Sources/SessionWeb/Navigation/SessionBrowserState.swift
